Make resowners more easily extensible.
Instead of having a separate array/hash for each resource kind, use a
single array and hash to hold all kinds of resources. This makes it
possible to introduce new resource "kinds" without having to modify the
ResourceOwnerData struct. In particular, this makes it possible for
extensions to register custom resource kinds.

The old approach was to have a small array of resources of each kind, and
if it fills up, switch to a hash table. The new approach also uses an
array and a hash, but now the array and a hash are used at the same time.
The array is used to hold the recently added resources, and when it fills
up, they are moved to the hash. This keeps the access to recent entries
fast, even when there are a lot of long-held resources.

All the resource-specific ResourceOwnerEnlarge*(), ResourceOwnerRemember*(),
and ResourceOwnerForget*() functions have been replaced with three generic
functions that take resource kind as argument. For convenience, we still
define resource-specific wrapper macros around the generic functions, with
the same old names, but they are now defined in the source files that use
those resource kinds.

Each resource kind specifies a release priority, and
ResourceOwnerReleaseAll releases the resources in priority order. To
make that possible, this restricts what you can do between phases.
After calling ResourceOwnerRelease, you are no longer allowed to use
the resource owner to remember any more resources in it, or to forget
any previously remembered resources by calling ResourceOwnerForget.
There was one case where that was done previously. At subtransaction
commit, AtEOSubXact_Inval() would handle the invalidation messages and
call RelationFlushRelation(), which temporarily increased the
reference count on the relation being flushed. We now switch to the
parent subtransaction's resource owner before calling
AtEOSubXact_Inval(), so that there is a valid ResourceOwner to
temporarily hold that relcache reference.

Other end-of-xact routines make similar calls to AtEOXact_Inval()
between release phases, but I didn't see any regression test failures
from those, so I'm not sure if they could reach a codepath that needs
remembering extra resources.

Also, the release callback no longer needs to call ResourceOwnerForget
on the resource being released. ResourceOwnerRelease unregisters the
resource from the owner before calling the callback. That needed some
changes in bufmgr.c and some other files, where releasing the resources
previously always called ResourceOwnerForget.

Add tests for ResourceOwners in src/test/modules/test_resowner.

- XXX in llvmjit.c

Reviewed-by: Aleksander Alekseev, Michael Paquier, Julien Rouhaud, Kyotaro Horiguchi, Hayato Kuroda, Álvaro Herrera, Zhihong Yu
Discussion: https://www.postgresql.org/message-id/cbfabeb0-cd3c-e951-a572-19b365ed314d%40iki.fi