Reduce overhead of small block data in xlog records
We reserve 2 bits in the block ID header for how large the block's registered
data is: 0, 1 or 2 bytes, for 0, 1-255 and 256-UINT16_MAX bytes to the block,
respectively. This reduces the minimum size of any block-referencing xlog
record from 48 bytes to 46 bytes, and reduces the size of any multi-page heap
update record by at least 2 bytes.

It includes more infrastructure for future value-dependent encoding of
length fields, even though it is currently only used in one place.