Explicitly skip TAP tests under Meson if disabled
If the tap_tests option is disabled under Meson, the TAP tests are
currently not registered at all.  But this makes it harder to see what
is going one, why suddently there are fewer tests than before.

Instead, register the tests anyway but with a dummy command that marks
them as skipped.  That way, the total list and count of tests is
constant whether the option is enabled or not.